+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+- Adds support for regular expression-based custom headers and rewrites for Firebase Hosting. (#2391)
 - Fixes a bug with RTDB Rules hot reloading in the RTDB emulator. (#2371)
 - Changes default functions runtime to Node.js 10 for `firebase init`.
+- Adds import/export support to the Realtime Database emulator.
+- Adds application default credentials inside the Functions emulator (#1940)